Factor out most of file loading from MoorhenContainer.
